灰色关联度代码更改

+++ 数学建模/数学建模算法/数学建模算法4/灰色关联度分析/GRA.py
@@ -68,7 +68,7 @@ def ShowGRAHeatMap(DataFrame):
     # mask[np.triu_indices_from(mask)] = True
 
     with sns.axes_style("white"):
-        sns.heatmap(DataFrame,
+        sns.heatmap(DataFrame.astype(float),
                     cmap="YlGnBu",
                     annot=True,
                     mask=mask,
